Refactored drawer.jsx little bit - need to work on refactoring drawer later
