+++ src/javascript/app_2/components/elements/drawer/index.jsx
@@ -1,0 +1,5 @@
+export * from './drawer_header';
+export * from './drawer_item';
+export * from './drawer_items';
+// export * from './drawer';
+export * from './toggle_drawer';
