+++ src/javascript/app_2/pages/account/lost_password.jsx
@@ -1,10 +1,40 @@
 import React, { Component } from 'react';
 import { connect }          from '../../store/connect';
+import Button               from '../../components/form/button';
+import PasswordField       from '../../components/form/password_field';
+import { localize }            from '../../../_common/localize';
 
 class LostPassword extends Component {
+    state = {
+        password: ''
+    }
+
+    onChange = (password) => {
+        this.setState({password});
+        console.log(this.state.password);
+    }
+
     render() {
         return (
-            <div><p>hello world</p></div>
+            <div className='static_full'>
+                <h1>{localize('Password reset')}</h1>
+                <p id='password_reset_description'>{localize('To reset your password, enter the email address you used to create your account into the field below and click \'Reset password\'.')}</p>
+
+                <form id='frm_lost_password'>
+                    <PasswordField
+                        type='number'
+                        name='amount'
+                        value={this.state.password}
+                        onChange={(password)=>this.onChange(password)}
+                        is_currency
+                        prefix={'hi'}
+                        is_nativepicker={true}
+                    />
+                    <Button type='submit' msg_id='form_error' text={localize('Reset Password')} />
+                </form>
+
+                <p>{localize('If you don’t receive the email within the next few minutes, please check your junk/spam folder.')}</p>
+            </div>
         );
     }
 }
